commenting out some unused code
