+++ lib/nier_cli/nier_cli_fork_utils/fileTypeUtils_fork/dat/datHashGenerator.dart
@@ -1,4 +1,4 @@
-import '../../utils/utils.dart';
+import '../../utils/utils_fork.dart';
 
 // From https://github.com/xxk-i/DATrepacker
 class HashInfo {
