+++ lib/nier_cli/nier_cli_fork_utils/fileTypeUtils_fork/pak/pakExtractor.dart
