+++ lib/nier_cli/nier_cli_fork_utils/fileTypeUtils_fork/xml/xml_extension.dart
