+++ lib/nier_cli/nier_cli_fork_utils/fileTypeUtils_fork/yax/xmlToYax.dart
@@ -1,7 +1,7 @@
 import 'dart:io';
 import 'package:xml/xml.dart';
 
-import '../../utils/utils.dart';
+import '../../utils/utils_fork.dart';
 import '../utils/ByteDataWrapper.dart';
 
 int _getTagId(XmlElement tag) {
